pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-qa"
version = "0.1.0"
description = "Multi-agent QA system powered by LangGraph + GPT-4o that autonomously plans, writes, runs, and reports tests for APIs, UIs, and Python codebases."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3",
    "langchain-openai>=0.2",
    "langgraph>=0.2",
    "httpx>=0.27",
    "playwright>=1.44",
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "typer>=0.12",
    "rich>=13.7",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "jsonschema>=4.22",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4",
    "mypy>=1.10",
    "pytest-cov>=5.0",
]

[project.scripts]
agentic-qa = "agentic_qa.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
